Clean up buildout structure.
Remove solr 4.8/4.9 cfg.

Use buildout cache and make plone version configurable on travis.

Run different solr versions on travis.

Remove bootstrap.py and replace it with bootstrap-buildout.py.

Attempt to cache solr download on travis.

Fix solr-5.2.x.cfg.

No need to test against Py 2.6.

Pin setuptools and buildout to latest versions for Plone 4.3.x.

Increate flake8-max-complexity to 25.

Ignore skins dir for code analysis. Run code-analysis before tests on travis.

Move versions to separate file. Allow Plone 5 tests to fail.

Allow Solr 5.1.x to fail.
