+++ .travis.yml
@@ -1,20 +1,29 @@
 language: python
 python:
-  - "2.6"
-  - "2.7"
+  - 2.7
 cache:
+  pip: true
   directories:
-  - download-cache
+    - $HOME/buildout-cache
+    - $HOME/parts/solr-download
+env:
+  - PLONE_VERSION=4.3.x SOLR_VERSION=4.10.x
+#  - PLONE_VERSION=5.0.x SOLR_VERSION=4.10.x
+  - PLONE_VERSION=4.3.x SOLR_VERSION=5.1.x
+#  - PLONE_VERSION=5.0.x SOLR_VERSION=5.1.x
+matrix:
+  allow_failures:
+    - env: PLONE_VERSION=4.3.x SOLR_VERSION=5.1.x
+before_install:
+  - mkdir -p $HOME/buildout-cache/{eggs,downloads,solr}
 install:
-  - mkdir -p download-cache
-  - mkdir -p buildout-cache/eggs
-  - mkdir -p buildout-cache/downloads
-  - pip install -U setuptools==1.4.2
-  - python bootstrap.py -c travis.cfg
+  - sed -ie "s#plone-x.x.x.cfg#plone-$PLONE_VERSION.cfg#" travis.cfg
+  - sed -ie "s#solr-x.x.x.cfg#solr-$SOLR_VERSION.cfg#" travis.cfg
+  - python bootstrap-buildout.py -c travis.cfg
   - bin/buildout -N -t 20 -c travis.cfg
 script:
-  - bin/test
   - bin/code-analysis
+  - bin/test
 after_success:
   - bin/createcoverage run bin/test
   - pip install -q coveralls
