+++ buildout.cfg
@@ -1,28 +1,7 @@
 [buildout]
 extends =
-    buildout/base.cfg
-    buildout/solr.cfg
-parts +=
-    autopep8
-    releaser
-    sphinxbuilder
-
-[solr-download]
-recipe = hexagonit.recipe.download
-strip-top-level-dir = true
-url = https://archive.apache.org/dist/lucene/solr/4.10.2/solr-4.10.2.tgz
-md5sum = a24f73f70e3fcf6aa8fda67444981f78
-
-[autopep8]
-recipe = zc.recipe.egg
-eggs = autopep8
-
-[releaser]
-recipe = zc.recipe.egg
-eggs =
-    zest.releaser
-    zest.pocompile
-
-[sphinxbuilder]
-recipe = collective.recipe.sphinxbuilder
-source = ${buildout:directory}/docs
+    base.cfg
+    plone-4.3.x.cfg
+    solr.cfg
+    solr-4.10.x.cfg
+    versions.cfg
