--- buildout/alltests.tmpl
@@ -1,20 +1,0 @@
-#!/usr/bin/env bash
-
-is_solr_up(){
-    http_code=`echo $(curl -s -o /dev/null -w "%{http_code}" "http://localhost:8983/solr/admin/ping")`
-    return `test $http_code = "200"`
-}
-
-wait_for_solr(){
-    while ! is_solr_up; do
-        sleep 3
-    done
-}
-
-${buildout:bin-directory}/solr-instance start
-wait_for_solr
-echo 'Solr is up. Running tests ...'
-${buildout:bin-directory}/test $*
-status=$?
-${buildout:bin-directory}/solr-instance stop
-exit $status
