--- buildout/base.cfg
@@ -1,37 +1,0 @@
-[buildout]
-extends = https://raw.github.com/collective/buildout.plonetest/master/plone-4.3.x.cfg
-package-name = collective.solr
-package-extras = [test]
-parts += test alltests code-analysis
-extensions = mr.developer
-sources-dir = extras
-
-[test]
-recipe = collective.xmltestreport
-eggs =
-    ${buildout:package-name} ${buildout:package-extras}
-
-[alltests]
-recipe = collective.recipe.template
-input = ${buildout:directory}/buildout/alltests.tmpl
-output = ${buildout:bin-directory}/alltests
-
-[code-analysis]
-recipe = plone.recipe.codeanalysis
-directory = ${buildout:directory}/src/collective/solr
-flake8-max-complexity = 50
-flake8-exclude = bootstrap.py,docs,*.egg.,skins
-#deprecated-aliases = True
-#utf8-header = True
-#imports = True
-#hasattr = True
-
-[versions]
-collective.recipe.solrinstance = 5.3.1
-collective.xmltestreport = 1.3.0
-coverage = 3.4
-zc.buildout = 1.7.1
-
-[sources]
-collective.indexing = git git@github.com:plone/collective.indexing.git
-plone.app.querystring = git git@github.com:plone/plone.app.querystring.git
