--- buildout/sphinx.cfg
@@ -1,7 +1,0 @@
-[buildout]
-parts += sphinxbuilder
-
-[sphinxbuilder]
-recipe = collective.recipe.sphinxbuilder
-source = ${buildout:directory}/docs/source
-build = ${buildout:directory}/docs
