+++ plone-4.3.x.cfg
@@ -1,0 +1,2 @@
+[buildout]
+extends = http://dist.plone.org/release/4.3.6/versions.cfg
