+++ plone-5.0.x.cfg
@@ -1,0 +1,2 @@
+[buildout]
+extends = http://dist.plone.org/release/5.0b2/versions.cfg
