+++ solr-4.10.x.cfg
@@ -1,6 +1,6 @@
 [buildout]
-extends = buildout.cfg
+parts += solr-download
 
 [solr-download]
-url = https://archive.apache.org/dist/lucene/solr/4.10.2/solr-4.10.2.tgz
-md5sum = a24f73f70e3fcf6aa8fda67444981f78
+url = https://archive.apache.org/dist/lucene/solr/4.10.4/solr-4.10.4.tgz
+md5sum = 8ae107a760b3fc1ec7358a303886ca06
