+++ solr-4.3.x.cfg
@@ -1,6 +1,2 @@
 [buildout]
-extends = buildout.cfg
-
-[solr-download]
-url = http://archive.apache.org/dist/lucene/solr/4.3.1/solr-4.3.1.tgz
-md5sum = a5d77cabd6f0e23d14d3918329f534bd
+extends = http://dist.plone.org/release/4.3.4/versions.cfg
