--- solr-4.8.x.cfg
@@ -1,6 +1,0 @@
-[buildout]
-extends = buildout.cfg
-
-[solr-download]
-url = http://archive.apache.org/dist/lucene/solr/4.8.0/solr-4.8.0.tgz
-md5sum = b12da59b9db9339e308e5c533b54331f
