--- solr-4.9.x.cfg
@@ -1,6 +1,0 @@
-[buildout]
-extends = buildout.cfg
-
-[solr-download]
-url = https://archive.apache.org/dist/lucene/solr/4.9.0/solr-4.9.0.tgz
-md5sum = 316f11ed8e81cf07ebfa6ad9443add09
