+++ solr-5.2.x.cfg
@@ -1,0 +1,3 @@
+[solr-download]
+url = https://archive.apache.org/dist/lucene/solr/5.2.1/solr-5.2.1-src.tgz
+md5sum = e022564b9a4b72c60e6e3c6bdf52654f
