+++ solr.cfg
@@ -3,15 +3,18 @@ parts +=
     solr-download
     solr-instance
 
+
 [settings]
 solr-host = 127.0.0.1
 solr-port = 8983
 solr-min-ram = 128M
 solr-max-ram = 256M
 
+
 [solr-download]
 recipe = hexagonit.recipe.download
 strip-top-level-dir = true
+
 
 [solr-instance]
 recipe = collective.recipe.solrinstance
