+++ travis.cfg
@@ -1,35 +1,19 @@
 [buildout]
 extends =
-    buildout/base.cfg
-    buildout/solr.cfg
+    base.cfg
+    plone-x.x.x.cfg
+    solr.cfg
+    solr-x.x.x.cfg
+    versions.cfg
 parts +=
-    download
-    install
     code-analysis
-    createcoverage
-eggs-directory = buildout-cache/eggs
-download-cache = download-cache
 
-[download]
-recipe = hexagonit.recipe.download
-url = https://launchpad.net/plone/4.3/4.3.4/+download/Plone-4.3.4-UnifiedInstaller.tgz
-
-[install]
-recipe = collective.recipe.cmd
-on_install = true
-cmds = tar jxvf ${download:location}/Plone-4.3.4-UnifiedInstaller/packages/buildout-cache.tar.bz2 1>/dev/null
-
-[solr-download]
-url = https://archive.apache.org/dist/lucene/solr/4.10.2/solr-4.10.2.tgz
-md5sum = a24f73f70e3fcf6aa8fda67444981f78
+# caches, see also .travis.yaml
+# one should not depend on '/home/travis' but it seems stable in containers.
+eggs-directory = /home/travis/buildout-cache/eggs
+download-cache = /home/travis/buildout-cache/downloads
 
 [code-analysis]
 recipe = plone.recipe.codeanalysis
-return-status-codes = True
-
-[createcoverage]
-recipe = zc.recipe.egg
-eggs = createcoverage
-
-[versions]
-zc.buildout = 1.6.3
+pre-commit-hook = False
+# return-status-codes = True
