+++ versions.cfg
@@ -1,0 +1,17 @@
+[versions]
+
+# Buildout
+setuptools = 18.0.1
+zc.buildout = 2.3.1
+
+# Code Analysis
+coverage = 3.7.1
+flake8 = 2.4.1
+pep8 = 1.5.7
+plone.recipe.codeanalysis = 2.0b1
+
+# Solr
+collective.recipe.solrinstance = 5.3.2
+
+# Misc
+zope.interface = 4.0.5
